# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "2.0.0",
    "title": "Volcanoes around the world v2",
    "description": "This API has been created to support assignment three in CAB230, Web Computing. It exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
    "schemas": {
      "MissingCountryParameterVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Country is a required query parameter."
          }
        }
      },
      "InvalidParametersVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Invalid query parameters. Only country and populatedWithin are permitted."
          }
        }
      },
...
      "VolcanoIdNotFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Volcano with ID: 99999 not found."
          }
        }
      },
      "InvalidPopulatedWithinParameterVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Invalid value for populatedWithin. Only: 5km,10km,30km,100km are permitted."
          }
        }
      },
      "InvalidParametersVolcano": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Invalid query parameters. Query parameters are not permitted."
          }
        }
      },
      "MissingAuthHeader": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Authorization header ('Bearer token') not found"
          }
        }
      },
      "TokenExpired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "JWT token has expired"
          }
        }
      },
      "InvalidJWT": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Invalid JWT token"
          }
        }
      },
      "MalformedAuthHeader": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "Authorization header is malformed"
          }
        }
      },
...
    }
  },
  "paths": {
    "/countries": {
      "get": {
        "tags": ["Data"],
        "description": "Returns a list of all countries that are associated with one or more volcanoes, ordered alphabetically.",
        "responses": {
          "200": {
            "description": "An array of countries. An example of the first five elements in the array is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Algeria",
                    "Antarctica",
                    "Argentina",
                    "Armenia",
                    "Australia"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid query parameters. Query parameters are not permitted."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/volcanoes": {
      "get": {
        "tags": ["Data"],
        "description": "Returns a list of volcanoes that are associated with the queried country. The country query parameter is required. The list can optionally be filtered by using the populatedWithin query parameter. This will return a filtered list of volcanoes that have at least one person living within the provided radius.",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "description": "Name of country",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populatedWithin",
            "in": "query",
            "description": "Distance within X km of a volcano",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["5km", "10km", "30km", "100km"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of objects containing id, name, country, region and subregion properties. An example of one object in the array is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": "1"
                      },
                      "name": {
                        "type": "string",
                        "example": "Abu"
                      },
                      "country": {
                        "type": "string",
                        "example": "Japan"
                      },
                      "region": {
                        "type": "string",
                        "example": "Japan, Taiwan, Marianas"
                      },
                      "subregion": {
                        "type": "string",
                        "example": "Honshu"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing country query parameter or query parameters invalid. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingCountryParameterVolcanoes"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidParametersVolcanoes"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidPopulatedWithinParameterVolcanoes"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/volcano/{id}": {
      "get": {
        "tags": ["Data"],
        "description": "Returns an object containing data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. To test this using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in. The path parameter (id) is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the volcano",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object containing id, name, country, region, subregion, last eruption, summit, elevation, latitude and longitude data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. An example of one object (with an authenticated request) is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Abu"
                    },
                    "country": {
                      "type": "string",
                      "example": "Japan"
                    },
                    "region": {
                      "type": "string",
                      "example": "Japan, Taiwan, Marianas"
                    },
                    "subregion": {
                      "type": "string",
                      "example": "Honshu"
                    },
                    "last_eruption": {
                      "type": "string",
                      "example": "6850 BCE"
                    },
                    "summit": {
                      "type": "number",
                      "example": "641"
                    },
                    "elevation": {
                      "type": "number",
                      "example": "2103"
                    },
                    "latitude": {
                      "type": "string",
                      "example": "34.5000"
                    },
                    "longitude": {
                      "type": "string",
                      "example": "131.6000"
                    },
                    "population_5km": {
                      "type": "number",
                      "example": "3597"
                    },
                    "population_10km": {
                      "type": "number",
                      "example": "9594"
                    },
                    "population_30km": {
                      "type": "number",
                      "example": "117805"
                    },
                    "population_100km": {
                      "type": "number",
                      "example": "4071152"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidParametersVolcano"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidJWT"
                    },
                    {
                      "$ref": "#/components/schemas/TokenExpired"
                    },
                    {
                      "$ref": "#/components/schemas/MalformedAuthHeader"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Volcano ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/VolcanoIdNotFound"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/volcano/{id}/comments": {
      "get": {
        "tags" : ["Data"],
        "description": "Returns an array of comments for the queried volcano. The path parameter (id) is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the volcano",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses":{
          "200": {
            "description": "An array of objects containing id of the comment , id of the volcano, email of the user who wrote the comment, the rating the user gave the volcano and th comment of the user. An example of one object in the array is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "volcanoId": {
                        "type": "number",
                        "example": 1
                      },
                      "userEmail": {
                        "type": "string",
                        "example": "mike@gmail.com"
                      },
                      "rating": {
                        "type": "number",
                        "example": 5
                      },
                      "comment": {
                        "type": "string",
                        "example": "This volcano is beautiful."
                      }
                    }
                  }
                }
              }
            }
      
          },
          "400" : {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidParametersVolcano"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No comments found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "error": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "No comments found for volcano with ID 9999999."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/volcano/{id}/addcomment": {
      "post": {
        "tags": ["Data"],
        "description": "Adds a new comment to the queried volcano. The path parameter (id) is required. A request body containing the rating and comment must be sent.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the volcano",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "An object containing the rating and the comment the user gave the volcano.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string",
                    "example": "This volcano is beautiful."
                  },
                  "rating": {
                    "type": "number",
                    "example": "5"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment added successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, both comment and rating are required"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already commented",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "You have already commented on this volcano."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingAuthHeader"
                    },
                    {
                      "$ref": "#/components/schemas/TokenExpired"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidJWT"
                    },
                    {
                      "$ref": "#/components/schemas/MalformedAuthHeader"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Volcano ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/VolcanoIdNotFound"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}